Add explorer dir list

+++ app/src/main/java/de/wollis_page/gibsonos/dto/desktop/Item.kt
@@ -7,6 +7,6 @@ class Item(
     var text: String,
     var icon: String,
     var thumb: String,
-    var customIcon: Int,
+    var customIcon: Long,
     //var params: LinkedHashMap<String, Any>
 )
